new pull

+++ server/api/v1/exa_customer.go
@@ -7,6 +7,7 @@ import (
 	"gin-vue-admin/model/request"
 	resp "gin-vue-admin/model/response"
 	"gin-vue-admin/service"
+	"gin-vue-admin/utils"
 	"github.com/gin-gonic/gin"
 )
 
@@ -21,6 +22,15 @@ import (
 func CreateExaCustomer(c *gin.Context) {
 	var cu model.ExaCustomer
 	_ = c.ShouldBindJSON(&cu)
+	CustomerVerify := utils.Rules{
+		"CustomerName":      {utils.NotEmpty()},
+		"CustomerPhoneData": {utils.NotEmpty()},
+	}
+	CustomerVerifyErr := utils.Verify(cu, CustomerVerify)
+	if CustomerVerifyErr != nil {
+		response.FailWithMessage(CustomerVerifyErr.Error(), c)
+		return
+	}
 	claims, _ := c.Get("claims")
 	waitUse := claims.(*request.CustomClaims)
 	cu.SysUserID = waitUse.ID
@@ -44,6 +54,14 @@ func CreateExaCustomer(c *gin.Context) {
 func DeleteExaCustomer(c *gin.Context) {
 	var cu model.ExaCustomer
 	_ = c.ShouldBindJSON(&cu)
+	CustomerVerify := utils.Rules{
+		"ID": {utils.NotEmpty()},
+	}
+	CustomerVerifyErr := utils.Verify(cu.Model, CustomerVerify)
+	if CustomerVerifyErr != nil {
+		response.FailWithMessage(CustomerVerifyErr.Error(), c)
+		return
+	}
 	err := service.DeleteExaCustomer(cu)
 	if err != nil {
 		response.FailWithMessage(fmt.Sprintf("删除失败：%v", err), c)
@@ -63,6 +81,23 @@ func DeleteExaCustomer(c *gin.Context) {
 func UpdateExaCustomer(c *gin.Context) {
 	var cu model.ExaCustomer
 	_ = c.ShouldBindJSON(&cu)
+	IdCustomerVerify := utils.Rules{
+		"ID": {utils.NotEmpty()},
+	}
+	IdCustomerVerifyErr := utils.Verify(cu.Model, IdCustomerVerify)
+	if IdCustomerVerifyErr != nil {
+		response.FailWithMessage(IdCustomerVerifyErr.Error(), c)
+		return
+	}
+	CustomerVerify := utils.Rules{
+		"CustomerName":      {utils.NotEmpty()},
+		"CustomerPhoneData": {utils.NotEmpty()},
+	}
+	CustomerVerifyErr := utils.Verify(cu, CustomerVerify)
+	if CustomerVerifyErr != nil {
+		response.FailWithMessage(CustomerVerifyErr.Error(), c)
+		return
+	}
 	err := service.UpdateExaCustomer(&cu)
 	if err != nil {
 		response.FailWithMessage(fmt.Sprintf("更新失败：%v", err), c)
@@ -82,6 +117,14 @@ func UpdateExaCustomer(c *gin.Context) {
 func GetExaCustomer(c *gin.Context) {
 	var cu model.ExaCustomer
 	_ = c.ShouldBindQuery(&cu)
+	IdCustomerVerify := utils.Rules{
+		"ID": {utils.NotEmpty()},
+	}
+	IdCustomerVerifyErr := utils.Verify(cu.Model, IdCustomerVerify)
+	if IdCustomerVerifyErr != nil {
+		response.FailWithMessage(IdCustomerVerifyErr.Error(), c)
+		return
+	}
 	err, customer := service.GetExaCustomer(cu.ID)
 	if err != nil {
 		response.FailWithMessage(fmt.Sprintf("获取失败：%v", err), c)
@@ -103,6 +146,11 @@ func GetExaCustomerList(c *gin.Context) {
 	waitUse := claims.(*request.CustomClaims)
 	var pageInfo request.PageInfo
 	_ = c.ShouldBindQuery(&pageInfo)
+	PageVerifyErr := utils.Verify(pageInfo, utils.CustomizeMap["PageVerify"])
+	if PageVerifyErr != nil {
+		response.FailWithMessage(PageVerifyErr.Error(), c)
+		return
+	}
 	err, customerList, total := service.GetCustomerInfoList(waitUse.AuthorityId, pageInfo)
 	if err != nil {
 		response.FailWithMessage(fmt.Sprintf("获取失败：%v", err), c)
